fix typo | move function to factory @crashillo
